move send function to nases service
