+++ nest-forms-backend/src/nases/nases.module.ts
@@ -12,7 +12,6 @@ import FormValidatorRegistryModule from '../form-validator-registry/form-validat
 import FormsHelper from '../forms/forms.helper'
 import FormsModule from '../forms/forms.module'
 import { FormsV2Module } from '../forms-v2/forms-v2.module'
-import NasesConsumerModule from '../nases-consumer/nases-consumer.module'
 import PrismaModule from '../prisma/prisma.module'
 import RabbitmqClientModule from '../rabbitmq-client/rabbitmq-client.module'
 import ScannerClientService from '../scanner-client/scanner-client.service'
@@ -30,7 +29,6 @@ import NasesUtilsService from './utils-services/tokens.nases.service'
     FormsModule,
     RabbitmqClientModule,
     FilesModule,
-    NasesConsumerModule,
     ConvertModule,
     TaxModule,
     ConvertPdfModule,
